Implemented deterministic physics simulation
Now physics is update at fixed rate and delta time - once every 1 / 120 of a second.

+++ Omniforce/src/Core/Input/Input.h
@@ -24,7 +24,7 @@ namespace Omni
 		static void LockAndHideMouse(const std::string& window_tag = "main") { return s_Instance->Impl_LockAndHideMouse(window_tag); }
 		static void ReleaseAndShowMouse(const std::string& window_tag = "main") { return s_Instance->Impl_ReleaseAndShowMouse(window_tag); }
 
-		static float32 Time() { return s_Instance->Impl_Time(); }
+		static float64 Time() { return s_Instance->Impl_Time(); }
 
 	protected:
 		Input() {};
@@ -36,7 +36,7 @@ namespace Omni
 		virtual vec2<int32>	Impl_MousePosition(const std::string& window_tag) const = 0;
 		virtual void		Impl_LockAndHideMouse(const std::string& window_tag) = 0;
 		virtual void		Impl_ReleaseAndShowMouse(const std::string& window_tag) = 0;
-		virtual float32		Impl_Time() = 0;
+		virtual float64		Impl_Time() = 0;
 
 	protected:
 		static Input* s_Instance;
