+++ Omniforce/src/Platform/Win64/Win64Input.cpp
@@ -50,7 +50,7 @@ namespace Omni {
 		glfwSetInputMode((GLFWwindow*)Application::Get()->GetWindow(window_tag)->Raw(), GLFW_CURSOR, GLFW_CURSOR_NORMAL);
 	}
 
-	float32 Win64Input::Impl_Time()
+	float64 Win64Input::Impl_Time()
 	{
 		return glfwGetTime();
 	}
